Split cache API and ehcache implementation into separate modules (#7171)
