+++ documentation/manual/releases/release26/migration26/CacheMigration26.md
@@ -1,12 +1,30 @@
 # Cache APIs Migration
 
+## New packages
+
+Now `cache` has been split into a `cacheApi` component with just the API, and `ehcache` that contains the Ehcache implementation. If you are using the default Ehcache implementation, simply change `cache` to `ehcache` in your `build.sbt`:
+
+```
+libraryDependencies ++= Seq(
+    ehcache
+)
+```
+
+If you are defining a custom cache API, or are writing a cache implementation module, you can just depend on the API:
+
+```
+libraryDependencies ++= Seq(
+    cacheApi
+)
+```
+
 ## Removed APIs
 
 The deprecated Java class `play.cache.Cache` was removed and you now must inject an `play.cache.SyncCacheApi` or `play.cache.AsyncCacheApi`.
 
 ## New Sync and Async Cache APIs
 
-The Cache API has been rewritten to have a synchronous and an asynchronous version. The old APIs will still work but they are now deprecated. 
+The Cache API has been rewritten to have a synchronous and an asynchronous version. The old APIs will still work but they are now deprecated.
 
 ### Java API
 
@@ -55,7 +73,7 @@ class Application @Inject() (cache: SyncCacheApi) extends Controller {
 ```
 
 Basically the same for `play.api.cache.AsyncCacheApi`:
- 
+
 ```scala
 class Application @Inject() (cache: AsyncCacheApi) extends Controller {
