+++ framework/src/play-ehcache/src/main/resources/ehcache-default.xml
