+++ framework/src/play-ehcache/src/main/resources/reference.conf
