+++ framework/src/play-ehcache/src/test/resources/ehcache-alternate.xml
