+++ framework/src/play-ehcache/src/test/resources/logback-test.xml
@@ -1,0 +1,20 @@
+<!--
+  ~ Copyright (C) 2009-2017 Lightbend Inc. <https://www.lightbend.com>
+  -->
+<configuration>
+    <!-- Suppress logback complaining about multiple logback-test.xml files -->
+    <statusListener class="ch.qos.logback.core.status.NopStatusListener" />
+
+    <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
+        <encoder>
+            <!-- We use short exception stack trace logging to limit output for travis. -->
+            <!-- Change to full if you need to do further debugging, but never commit that. -->
+            <pattern>%level %logger{15} - %message%n%ex{short}</pattern>
+        </encoder>
+    </appender>
+
+    <root level="WARN">
+        <appender-ref ref="STDOUT" />
+    </root>
+
+</configuration>
