+++ framework/src/play-ehcache/src/test/scala/play/api/cache/CachedSpec.scala
